Move plan.Diff into applied package
