+++ internal/cmd/agent/deployer/applied/diff.go
@@ -1,4 +1,4 @@
-package plan
+package applied
 
 import (
 	"encoding/json"
